Code updates
Added bin2dec.asm to test conversion from a binary number to decimal number (dividing it by 10 in order to find the remainders and print them in reverse sequence in order to print the base10 'human-readable' number. Also created a shorter version of lcd6 (including a number of routines to drive the LCD) called lcd7, that I used to describe JSR decoding issues that I found on nectivio 65C02-Tool. Next step is to go back to lcd6 and improve speed of LCD scrolling routine; see if I can optimize it and make it less complex.